services/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
asyncio-mqtt==0.16.1
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0
prometheus-client==0.19.0
structlog==23.2.0
openai==1.3.0
python-dotenv==1.0.0
chromadb==0.4.18
tiktoken==0.5.2